Make string reads have word-wide characters
The current Rust print instructions and entire Java implementation
depend on characters being stored in 4- or 8-long words, as opposed to
single bytes. As such, I have modified the string reads to 0-pad
their byte vector prior to adding the entire vector as a single
transformation.

In the future, we should rewrite this implementation to store the
characters in byte-wide slots to optimize our memory usage.