Create ReaderCell
Since we need access to the standard input reader both on the side of
the input and the consumption. As such, we need to use the interior
mutability model, in which a reader cell holds access to the reader, and
distributes it to at most one referrer at a time.

This reader cell now replaces the `ReaderBox` throughout the code, so I
had to update any code that referred to it.

Additionally, I added the io::Write trait to the Reader trait, since
it is the typical way to explain that a buffer is expandable within an
element. I modified some code in the Tauri app to further work with
this.