Add directory flag to read_dir
Currently, our file system function to read a directory fails to inform
the front-end whether a path leads to a file or a directory. As such, I
have added a boolean to each output with this information.

Also, renamed copy to copy_file for consistency, clarity, and
compatibility with the initial interface.