Restructure tests
In Rust, it is preferred to write tests in the same file as the code it
is testing, as this allows for increased visibility and a localization
of logic. Since these factors benefit our project, I have moved the
tests around.

Now to run tests, you must use the `cargo test --all --lib` command.

Additionally, since visibility into the instruction files is no longer
necessary, all instruction macros and functions are hidden in the
module, and the instruction registration methods act as an interface for
them.